# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "YGG API",
    "description": "API de análisis de partidas de League of Legends",
    "version": "2.0.0"
  },
  "paths": {
...
    "/matches/snapshot/{snapshot_id}": {
      "get": {
        "tags": [
          "matches"
        ],
        "summary": "List Matches",
        "description": "Partidas de un snapshot, de la más reciente a la más antigua (solo BD).",
        "operationId": "matches_list_matches",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Elementos por página",
              "default": 50,
              "title": "Limit"
            },
            "description": "Elementos por página"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Elementos a saltar",
              "default": 0,
              "title": "Offset"
            },
            "description": "Elementos a saltar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_MatchResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/snapshots/player/{player_id}": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "List Snapshots",
        "description": "Snapshots de un jugador, del más reciente al más antiguo.",
        "operationId": "snapshots_list_snapshots",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "player_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Player Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Elementos por página",
              "default": 50,
              "title": "Limit"
            },
            "description": "Elementos por página"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Elementos a saltar",
              "default": 0,
              "title": "Offset"
            },
            "description": "Elementos a saltar"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SnapshotResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/snapshots/": {
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Create Snapshot",
        "description": "Lanza el análisis de un período. Devuelve un job_id; el análisis corre en segundo plano.",
        "operationId": "snapshots_create_snapshot",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotJobResponse"
                }
              }
            }
          },
          "429": {
            "description": "Too many analyses requested"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/snapshots/jobs/{job_id}": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "Get Job Status",
        "description": "Estado de un análisis. Solo lo ve quien lo lanzó; para cualquier otro es un 404.",
        "operationId": "snapshots_get_job_status",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/snapshots/{snapshot_id}": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "Get Snapshot",
        "operationId": "snapshots_get_snapshot",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "snapshots"
        ],
        "summary": "Remove Snapshot",
        "description": "Elimina un snapshot; las partidas se conservan para otros snapshots.",
        "operationId": "snapshots_remove_snapshot",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/snapshots/{snapshot_id}/description": {
      "patch": {
        "tags": [
          "snapshots"
        ],
        "summary": "Edit Description",
        "operationId": "snapshots_edit_description",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotDescriptionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/snapshots/{snapshot_id}/notes": {
      "patch": {
        "tags": [
          "snapshots"
        ],
        "summary": "Edit Notes",
        "operationId": "snapshots_edit_notes",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SnapshotNotesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/snapshots/{snapshot_id}/dashboard": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "Read Snapshot Dashboard",
        "description": "Dashboard del snapshot (cacheado). `compare_*` superpone en vivo el radar de otra cuenta.",
        "operationId": "snapshots_read_snapshot_dashboard",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Snapshot Id"
            }
          },
          {
            "name": "compare_game_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare Game Name"
            }
          },
          {
            "name": "compare_tag_line",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare Tag Line"
            }
          },
          {
            "name": "compare_region",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare Region"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotDashboardResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChampionStatsResponse": {
        "properties": {
          "champion_name": {
            "type": "string",
            "title": "Champion Name"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "win_rate": {
            "type": "number",
            "title": "Win Rate"
          },
          "icon_url": {
            "type": "string",
            "title": "Icon Url"
          }
        },
        "type": "object",
        "required": [
          "champion_name",
          "games_played",
          "win_rate",
          "icon_url"
        ],
        "title": "ChampionStatsResponse"
      },
...
      "DeathsByPhase": {
        "properties": {
          "early_deaths": {
            "type": "integer",
            "title": "Early Deaths"
          },
          "mid_deaths": {
            "type": "integer",
            "title": "Mid Deaths"
          },
          "late_deaths": {
            "type": "integer",
            "title": "Late Deaths"
          }
        },
        "type": "object",
        "required": [
          "early_deaths",
          "mid_deaths",
          "late_deaths"
        ],
        "title": "DeathsByPhase"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_MatchResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MatchResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MatchResponse]"
      },
...
      "Page_SnapshotResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SnapshotResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[SnapshotResponse]"
      },
...
      "RoleAverageMetric": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "threshold": {
            "type": "number",
            "title": "Threshold"
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "value",
          "status",
          "threshold"
        ],
        "title": "RoleAverageMetric"
      },
...
      "SnapshotCreate": {
        "properties": {
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "date_from": {
            "type": "integer",
            "title": "Date From"
          },
          "date_to": {
            "type": "integer",
            "title": "Date To"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "player_id",
          "date_from",
          "date_to"
        ],
        "title": "SnapshotCreate",
        "description": "Lo que recibe la API al lanzar un análisis."
      },
      "SnapshotDashboardResponse": {
        "properties": {
          "snapshot_id": {
            "type": "integer",
            "title": "Snapshot Id"
          },
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "player_name": {
            "type": "string",
            "title": "Player Name"
          },
          "date_from": {
            "type": "string",
            "format": "date-time",
            "title": "Date From"
          },
          "date_to": {
            "type": "string",
            "format": "date-time",
            "title": "Date To"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "default": ""
          },
          "active_role": {
            "type": "string",
            "title": "Active Role"
          },
          "games_played": {
            "type": "integer",
            "title": "Games Played"
          },
          "role_averages": {
            "items": {
              "$ref": "#/components/schemas/RoleAverageMetric"
            },
            "type": "array",
            "title": "Role Averages"
          },
          "played_champions": {
            "items": {
              "$ref": "#/components/schemas/ChampionStatsResponse"
            },
            "type": "array",
            "title": "Played Champions"
          },
          "deaths_by_phase": {
            "$ref": "#/components/schemas/DeathsByPhase"
          },
          "performance_trends": {
            "items": {
              "$ref": "#/components/schemas/TrendPoint"
            },
            "type": "array",
            "title": "Performance Trends"
          },
          "radar_data": {
            "$ref": "#/components/schemas/RadarChartData"
          }
        },
        "type": "object",
        "required": [
          "snapshot_id",
          "player_id",
          "player_name",
          "date_from",
          "date_to",
          "active_role",
          "games_played",
          "role_averages",
          "played_champions",
          "deaths_by_phase",
          "performance_trends",
          "radar_data"
        ],
        "title": "SnapshotDashboardResponse",
        "description": "Dashboard de un snapshot.\n\nNo incluye la lista de partidas (P9): se piden paginadas en\nGET /matches/snapshot/{snapshot_id}. Así la respuesta cabe en caché y no\ncrece con el número de partidas."
      },
      "SnapshotDescriptionUpdate": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "description"
        ],
        "title": "SnapshotDescriptionUpdate",
        "description": "Para editar la descripción de un snapshot."
      },
      "SnapshotJobResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "processing"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "SnapshotJobResponse",
        "description": "Respuesta inmediata al crear un snapshot (tarea en background)."
      },
      "SnapshotJobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "title": "Progress",
            "default": 0
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "SnapshotJobStatus",
        "description": "Estado de un job en curso."
      },
      "SnapshotNotesUpdate": {
        "properties": {
          "notes": {
            "type": "string",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "notes"
        ],
        "title": "SnapshotNotesUpdate",
        "description": "Para editar las notas inline desde la UI."
      },
      "SnapshotResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "player_id": {
            "type": "integer",
            "title": "Player Id"
          },
          "date_from": {
            "type": "string",
            "format": "date-time",
            "title": "Date From"
          },
          "date_to": {
            "type": "string",
            "format": "date-time",
            "title": "Date To"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "default": ""
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes",
            "default": ""
          },
          "match_count": {
            "type": "integer",
            "title": "Match Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "player_id",
          "date_from",
          "date_to"
        ],
        "title": "SnapshotResponse",
        "description": "Lo que devuelve la API al listar snapshots."
      },
...
      "TrendPoint": {
        "properties": {
          "game_num": {
            "type": "integer",
            "title": "Game Num"
          },
          "match_id": {
            "type": "string",
            "title": "Match Id"
          },
          "creation_time": {
            "type": "string",
            "format": "date-time",
            "title": "Creation Time"
          },
          "champion": {
            "type": "string",
            "title": "Champion"
          },
          "win": {
            "type": "boolean",
            "title": "Win"
          },
          "kda": {
            "type": "number",
            "title": "Kda"
          },
          "cs_per_min": {
            "type": "number",
            "title": "Cs Per Min"
          },
          "gold_per_min": {
            "type": "number",
            "title": "Gold Per Min"
          },
          "vision_per_min": {
            "type": "number",
            "title": "Vision Per Min"
          },
          "kda_moving_avg": {
            "type": "number",
            "title": "Kda Moving Avg"
          },
          "cs_moving_avg": {
            "type": "number",
            "title": "Cs Moving Avg"
          },
          "gold_moving_avg": {
            "type": "number",
            "title": "Gold Moving Avg"
          },
          "vision_moving_avg": {
            "type": "number",
            "title": "Vision Moving Avg"
          }
        },
        "type": "object",
        "required": [
          "game_num",
          "match_id",
          "creation_time",
          "champion",
          "win",
          "kda",
          "cs_per_min",
          "gold_per_min",
          "vision_per_min",
          "kda_moving_avg",
          "cs_moving_avg",
          "gold_moving_avg",
          "vision_moving_avg"
        ],
        "title": "TrendPoint"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}